disable the keycloak postgresql subchart
Because the version of the postgresql subchart used in the codecentric chart
is no longer available, we now default to an "external".

However, to keep the 'out-of-the-box' experience as smooth as possible,
we set all the keycloak DB_*  env values to the corresponding defaults of the
postgresql subchart.

This means that, if you enable both our postgresql subchart and the keycloak subchart
with their default values, the keycloak subchart will now use the same database
as the FHIR Server (whereas previously it deployed a separate one).

Signed-off-by: Lee Surprenant <lmsurpre@merative.com>